SettingsMap: make SettingValue's lazy type cache safe for concurrent reads
The per-model buffer styles render member models in parallel while all
bodies read the same SettingsMap.  SettingValue cached its parsed value
in a mutable curType + union, so two threads' first reads could publish
the tag before the value (or tear across a re-parse) and hand an effect
a garbage parameter for one model for one frame.  Pack the tag and the
32-bit value pattern into a single atomic word — concurrent first-reads
both parse and store the identical word.  Doubles don't fit the packed
word and are rare in render paths, so getDouble just parses per call.

+++ src-core/effects/OnEffect.h
@@ -42,7 +42,7 @@ public:
     static int sStartDefault;
     static int sEndDefault;
     static bool sShimmerDefault;
-    static double sCyclesDefault;
+    static float sCyclesDefault;
     static int sTransparencyDefault;
     static int sTransparencyMin;
     static int sTransparencyMax;
